+++ stories/Button.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import Button from '../src/components/Button';
 
-const label1 = "Hello";
+const label1 = 'Hello';
 const label2 = 'Click Me';
 
 storiesOf('Button', module)