+++ stories/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {storiesOf} from '@storybook/react';
+import { storiesOf } from '@storybook/react';
 import Footer from '../src/components/Footer';
 
 storiesOf('Footer', module)